Task 39. React Router
Подключи react-router-dom к своему приложению.

Проведи рефакторинг приложения, создай меню и дополнительные компоненты для разделов "Обо мне" и "Контакты".

Подключи и настрой роутинг в приложении.
