+++ src/components/About/About.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function About() {
+    return (
+        <div>
+            Обо мне
+        </div>
+    )
+};
+
+export default About;
