+++ src/components/App/App.module.css
@@ -1,21 +1,41 @@
-@import url('https://fonts.googleapis.com/css2?family=Jost:wght@700&display=swap');
 body {
     background-color: #212121;
 }
 
-.todo {
-    padding: 0;
-    margin: auto;
-    width: 50%;
+.menu {
     background-color: #333333;
     border: 1px solid #333333;
     border-radius: 10px;
+    margin-right: 10px;
 }
 
-.todo__title {
-    font-size: 45px;
-    margin: 0;
-    text-align: center;
-    color: white;
-    font-family: 'Jost', sans-serif;
+.element {
+    background-color: #333333;
+    border: 1px solid #333333;
+    border-radius: 10px;
+    color: #ffffff;
+    padding: 10px 10px 0px 10px;
+    width: 100%;
+}
+
+.wrap {
+    display: flex;
+    justify-content: space-between;
+    margin: auto;
+    width: 60%;
+}
+
+.menu__link {
+    text-decoration: none;
+    color: #fff;
+}
+
+.menu__link:hover {
+    color: darkviolet;
+    background-color: black;
+}
+
+.link__text:hover {
+    background-color: #212121;
+    border-radius: 10px;
 }
