+++ src/components/Contacts/Contacts.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function Contacts() {
+    return (
+        <div>
+            Контакты
+        </div>
+    );
+};
+
+export default Contacts;
